+++ src/services/event_driven/domain/types.py
@@ -6,7 +6,7 @@ from models.event import EventRecordAttribute
 
 from .models import EventRecord
 
-VendorKind = Literal["AWS", "MAESTRO"]
+VendorKind = Literal["AWS", "MAESTRO", "SRE_K8S_AGENT"]
 
 # Type aliases
 RegionNameType = str
@@ -18,3 +18,4 @@ EventNameType = str
 
 ESourceENameRulesMap = dict[EventSourceType, dict[EventNameType, list[RuleNameType]]]
 EventGenericRecord = EventRecordAttribute | EventRecord
+K8sServiceRulesMap = dict[str, list[RuleNameType]]